+++ fabric-events-interaction-v0/src/main/resources/fabric-events-interaction-v0.mixins.json
@@ -3,11 +3,11 @@
   "package": "net.fabricmc.fabric.mixin.event.interaction",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "MixinBlock",
+    "BlockMixin",
     "MixinServerPlayerEntity",
     "MixinServerPlayerInteractionManager",
     "MixinServerPlayNetworkHandler",
-    "MixinWorld"
+    "WorldMixin"
   ],
   "client": [
     "MixinClientPlayerInteractionManager",
